Refactor new endpoints and solved merge issues
