+++ index.js
@@ -22,6 +22,8 @@ const cancelRoute = require('./api/cancel-reservation');
 const returnRoute = require('./api/return');
 const listingRoute = require('./api/listing');
 const userRoute = require('./api/user');
+const loginRoute = require('./api/login');
+const registerRoute = require('./api/register');
 
 const sql = postgres({
   host: process.env.DB_HOST,
@@ -99,200 +101,11 @@ app.use('/api/item', listingRoute);
 // GET /api/user/:id endpoint
 app.use('/api', userRoute);
 
-app.post('/listing/:id', async (req, res) => {
-  const id = parseInt(req.params.id);
-  const user_id = parseInt(req.query.user_id);
-  const item_listing = req.query.item_listing;
-  let status;
+// POST /api/login endpoint
+app.use('/api', loginRoute);
 
-  if (isNaN(id) || isNaN(user_id) || !item_listing) {
-    res.status(400).send({
-      error: "BAD_REQUEST",
-      message: "Please provide a numeric ID, user ID, and all query parameters (user_id, item_listing)"
-    });
-  }
-
-  status = (item_listing === 'true');
-
-  try {
-    // select resource
-    const [resource] = await sql`
-      select owned_by, reservation_status 
-      from resources
-      where id = ${id} 
-    `;
-
-    // No resource found
-    if (!resource) {
-      return res.status(404).send({
-        error: API_RETURN_MESSAGES.ITEM_UNAVAILABLE,
-        id
-      });
-    } else if (resource.owned_by != user_id) { // user relisting the item must be its owner
-      return res.status(401).send({
-        error: API_RETURN_MESSAGES.UNAUTHORIZED,
-        message: "You are not authorized to perform this action"
-      });
-    } else if (resource.reservation_status != ITEM.LISTED && resource.reservation_status != ITEM.UNLISTED) { // the item cannot be reserved
-      return res.status(403).send({
-        error: API_RETURN_MESSAGES.RESERVED,
-        message: "Cannot alter the listing of an item that is currently reserved"
-      });
-    }
-
-    // update resource
-    const update_item_listing = await sql`
-      update resources
-      set reservation_status = ${(status ? ITEM.LISTED : ITEM.UNLISTED)}
-      where id = ${id}
-    `;
-
-    if (status) {
-      return res.status(200).send({
-        ok: API_RETURN_MESSAGES.RELISTED,
-        id
-      });
-
-    } else {
-      return res.status(200).send({
-        ok: API_RETURN_MESSAGES.UNLISTED,
-        id
-      });
-    }
-  }
-  catch (error) {
-    console.error(error);
-
-    res.status(500).send({
-      error: API_RETURN_MESSAGES.INTERNAL_SERVER_ERROR,
-      message: "Internal Server Error"
-    });
-  }
-});
-
-app.get('/api/user/:id', async (req, res) => {
-  const user_id = parseInt(req.params.id);
-
-  if (isNaN(user_id)) {
-    return res.status(400).send({
-      error: "BAD_REQUEST",
-      message: "please provide a numeric user ID"
-    });
-  }
-
-  try {
-    // Fetch the user
-    const [user] = await sql`
-      select id, name, email, phone_number
-      from users
-      where id = ${user_id};
-    `;
-
-    // If `!user`, a user under the given ID doesn't exist
-    if (!user) {
-      return res.status(404).send({
-        error: API_RETURN_MESSAGES.USER_NOT_FOUND,
-        id: user_id
-      });
-    }
-
-    res.status(200).send(user);
-
-  } catch (error) {
-    console.error(error);
-
-    res.status(500).send({
-      error: API_RETURN_MESSAGES.INTERNAL_SERVER_ERROR,
-      message: "Internal Server Error"
-    });
-  }
-});
-
-
-app.post('/api/login', async (req, res) => {
-  try {
-    const { email, password } = req.body;
-
-    if (email == undefined || password == undefined) {
-      return res.status(400).send({
-        error: "All_PARAMETER_REQUIRED"
-      });
-    }
-
-    const rows = await sql`SELECT * FROM users where email = ${email};`
-
-    if (rows.length === 0) {
-      return res.status(401).json({ error: 'INVALID_EMAIL_OR_PASSWORD' });
-    }
-
-    const user = rows[0];
-    const valid = await bcrypt.compare(password, user.password_hash);
-    if (!valid) {
-      return res.status(401).json({ error: 'INVALID_EMAIL_OR_PASSWORD' });
-    }
-
-    const token = jwt.sign({ id: user.id, email: user.email }, process.env.JWT_SECRET, {
-      expiresIn: process.env.JWT_EXPIRES_IN,
-    });
-    return res.status(200).json({ token, id: user.id });
-  } catch (err) {
-    console.error(err);
-    return res.status(500).json({ error: 'INTERNAL_SERVER_ERROR' });
-  }
-});
-
-
-// register new user
-app.post('/api/register', async (req, res) => {
-  var { name, email, phone_number, password } = req.body;
-
-  if (!email || !name || !password) {
-    res.status(400).json({ error: 'All_FIELDS_ARE_REQUIRED' });
-    return;
-  }
-
-  if (phone_number && phone_number.length != 10) {
-    res.status(400).json({ error: 'INVALID_PHONE_NUMBER' });
-    return;
-  } else if (phone_number == undefined) {
-    phone_number = ""
-  }
-
-  if (!emailRegex.test(email)) {
-    res.status(400).json({ error: 'INVALID_EMAIL' });
-    return;
-  }
-
-  try {
-    const existingUser = await sql`
-      SELECT * FROM users
-      WHERE email = ${email}
-      LIMIT 1
-    `;
-
-    if (existingUser.length > 0) {
-      res.status(400).json({
-        "error": "EMAIL_ALREADY_REGISTERED",
-        "email": email
-      })
-      return;
-    }
-
-    const hashedPassword = await bcrypt.hash(password, 10);
-    await sql`
-      INSERT INTO users (name,email,password_hash, phone_number)
-      VALUES (${name}, ${email}, ${hashedPassword}, ${phone_number})
-    `;
-
-    res.status(200).json({
-      ok: "REGISTERED_SUCCESFULLY"
-    });
-  } catch (err) {
-    console.error(err);
-    res.status(500).json({ error: "INTERNAL_SERVER_ERROR" })
-  }
-});
-
+// POST /api/register endpoint
+app.use('/api', registerRoute);
 
 // Start the webserver
 app.listen(port, () => {
